Review annotator policies for dataset settings (#2547)
For an annotator, dataset settings are visible in read-only mode.

This PR adds task dataset settings-related policies and uses them in
dataset settings-related actions.

Also, new tests are added to test this functionality.

---------

+++ src/argilla/server/services/datasets.py
@@ -30,7 +30,11 @@ from argilla.server.errors import (
     WrongTaskError,
 )
 from argilla.server.models import User, Workspace
-from argilla.server.policies import DatasetPolicy, is_authorized
+from argilla.server.policies import (
+    DatasetPolicy,
+    DatasetSettingsPolicy,
+    is_authorized,
+)
 from argilla.server.schemas.datasets import CreateDatasetRequest, Dataset
 
 
@@ -38,12 +42,8 @@ class ServiceBaseDataset(BaseDatasetDB):
     pass
 
 
-class ServiceBaseDatasetSettings(BaseDatasetSettingsDB):
-    pass
-
-
 ServiceDataset = TypeVar("ServiceDataset", bound=ServiceBaseDataset)
-ServiceDatasetSettings = TypeVar("ServiceDatasetSettings", bound=ServiceBaseDatasetSettings)
+ServiceDatasetSettings = TypeVar("ServiceDatasetSettings", bound=BaseDatasetSettingsDB)
 
 
 class DatasetsService:
@@ -226,6 +226,9 @@ class DatasetsService:
         settings = self.__dao__.load_settings(dataset=dataset, as_class=class_type)
         if not settings:
             raise EntityNotFoundError(name=dataset.name, type=class_type)
+
+        if not is_authorized(user, DatasetSettingsPolicy.list(dataset)):
+            raise ForbiddenOperationError("You don't have the necessary permissions to list settings for this dataset.")
         return class_type.parse_obj(settings.dict())
 
     def raw_dataset_update(self, dataset):
@@ -234,8 +237,16 @@ class DatasetsService:
     async def save_settings(
         self, user: User, dataset: ServiceDataset, settings: ServiceDatasetSettings
     ) -> ServiceDatasetSettings:
+        if not is_authorized(user, DatasetSettingsPolicy.save(dataset)):
+            raise ForbiddenOperationError("You don't have the necessary permissions to save settings for this dataset.")
+
         self.__dao__.save_settings(dataset=dataset, settings=settings)
         return settings
 
     async def delete_settings(self, user: User, dataset: ServiceDataset) -> None:
+        if not is_authorized(user, DatasetSettingsPolicy.delete(dataset)):
+            raise ForbiddenOperationError(
+                "You don't have the necessary permissions to delete settings for this dataset."
+            )
+
         self.__dao__.delete_settings(dataset=dataset)
